RELATED: BB-1855 color strategies refactoring
