+++ src/components/visualizations/chart/colorStrategies/test/helpers.ts
@@ -1,0 +1,12 @@
+// (C) 2020 GoodData Corporation
+import { IColorStrategy } from "../../colorFactory";
+
+export function getColorsFromStrategy(strategy: IColorStrategy): string[] {
+    const res: string[] = [];
+
+    for (let i = 0; i < strategy.getFullColorAssignment().length; i++) {
+        res.push(strategy.getColorByIndex(i));
+    }
+
+    return res;
+}
